Migrate to spectre.css lib (base size is 1rem=10px now)
Due to the way the font-size is defined on the html element in
spectre.css, some font-size and related definitions needed to change.

The introduction of spectre is mainly motivated by the need for a css
grid system and to help get a somewhat nice look without re-inventing
the wheel completely.